+++ src/TodoList.js
@@ -1,0 +1,28 @@
+import React from 'react'
+
+const todoList = [
+    {
+      id: 1,
+      title: "JavaScript Basic"
+    },
+    {
+      id: 2,
+      title: "Become a React.JS Developer"
+    },
+    {
+      id: 3,
+      title: "Take Node.JS Course"
+    }
+  ]
+
+const TodoList = () => {
+  return (
+    <div>
+        <ul>
+            {todoList.map((list) => (<li key={list.id}> {list.title} </li>))}
+        </ul>
+    </div>
+  )
+}
+
+export default TodoList
